Separate out optimisation_result
